rename package, move to app.py instead of __init__.py
