+++ MetadataManager/.env.example
