+++ MetadataManager/Dockerfile
@@ -10,7 +10,7 @@ WORKDIR /home/herop
 COPY --chown=herop:herop requirements.txt requirements.txt
 RUN pip3 --disable-pip-version-check install -U -r requirements.txt
 COPY --chown=herop:herop . .
-RUN pip3 install .
+RUN pip3 install . --disable-pip-version-check install
 
 EXPOSE 80
-CMD ["flask", "run", "-h", "0.0.0.0", "-p", "80"]
+CMD ["flask", "--app", "manager.app", "run", "-h", "0.0.0.0", "-p", "80"]
