+++ MetadataManager/manager/metadata/Aardvark/Metadata_Template.md
