+++ MetadataManager/manager/metadata/Aardvark/SocialVulnerabilityIndex.md
