+++ MetadataManager/manager/metadata/DataDictionaries/social-vulnerability-index.md
