+++ MetadataManager/manager/service/solr.py
