+++ MetadataManager/manager/static/style.css
