+++ README.md
@@ -17,13 +17,13 @@ A dev deploy will serve the app on Flask's default port (5000).
 ```
 python3 -m venv env
 source ./env/bin/activate
-pip install -e ./flask
+pip install -e ./MetadataManager
 ```
 
 Then
 
 ```
-flask run --debug
+flask --app MetadataManager.manager.app run --debug
 ```
 
 `--debug` will auto-reload the app whenever a file is changed (though it seems like changes to HTML files requires the app to be stopped and restarted...).
