+++ nginx/nginx.conf
@@ -7,7 +7,7 @@ http {
     sendfile on;
 
     upstream flask {
-        server flask;
+        server MetadataManager;
     }
 
     proxy_set_header   Host $host;
